Enhancing vibes, deduplicating, adding metaprogramming concepts
